+++ _site-theme-doceo/.kasar/theme/tasks/build.project.pages.js
@@ -244,7 +244,7 @@ function _convertWeb2HTML(webpages, menu) {
       const html = createPage(webpages, lang, page, content, _getNormalize(), params, menu);
 
       if (output) {
-        fse.outputFile(`${basedist}/${output}`, html, (err) => {
+        fse.outputFile(`${dist}/${output}`, html, (err) => {
           if (err) {
             throw new Error(err);
           }
@@ -395,7 +395,7 @@ function _convertDoc2HTML(website, menu, docpages, docmenu) {
       const html = createPage(website, lang, page, content, _getNormalize(), params, menu, docpages, cdocmenu);
 
       if (output) {
-        fse.outputFile(`${basedist}/${output}`, html, (err) => {
+        fse.outputFile(`${dist}/${output}`, html, (err) => {
           if (err) {
             throw new Error(err);
           }