+++ _site-theme-start/.kasar/theme/pages/main.js
@@ -40,7 +40,7 @@ const { product }    = config
 /**
  * Creates the web page in a virtual DOM and returns its HTML representation.
  *
- * @function (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, [arg9])
+ * @function (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9)
  * @public
  * @param {Object}          the website pages,
  * @param {Object}          the docsite pages,